+++ src/main/kotlin/com/example/reactornetdisk/repository/UserRepository.kt
@@ -4,7 +4,7 @@ import com.example.reactornetdisk.entity.User
 import org.springframework.data.repository.reactive.ReactiveCrudRepository
 import reactor.core.publisher.Mono
 
-interface UserRepository : ReactiveCrudRepository<User, Long> {
+interface UserRepository : ReactiveCrudRepository<User, Int> {
 
     fun findUserByUsername(username: String): Mono<User>
 }
